add deck generate to chebpoly; refactoring
